+++ lib/tapioca/compilers/symbol_table_compiler.rb
@@ -28,6 +28,7 @@ module Tapioca
         @bootstrap_symbols.each { |symbol| push_symbol(symbol) }
 
         @node_listeners = T.let([], T::Array[Gem::Listeners::Base])
+        @node_listeners << Gem::Listeners::Mixins.new(self)
         @node_listeners << Gem::Listeners::SorbetHelpers.new(self)
         @node_listeners << Gem::Listeners::SorbetEnums.new(self)
         @node_listeners << Gem::Listeners::SorbetProps.new(self)
@@ -209,7 +210,6 @@ module Tapioca
         # Compiling type variables must happen first to populate generic names
         compile_type_variables(tree, constant)
         compile_methods(tree, name, constant)
-        compile_mixins(tree, constant)
         compile_dynamic_mixins(tree, constant)
       end
 
@@ -321,55 +321,6 @@ module Tapioca
         push_symbol(name)
 
         "::#{name}"
-      end
-
-      sig { params(tree: RBI::Tree, constant: Module).void }
-      def compile_mixins(tree, constant)
-        singleton_class = singleton_class_of(constant)
-
-        interesting_ancestors = interesting_ancestors_of(constant)
-        interesting_singleton_class_ancestors = interesting_ancestors_of(singleton_class)
-
-        prepends = interesting_ancestors.take_while { |c| !are_equal?(constant, c) }
-        includes = interesting_ancestors.drop(prepends.size + 1)
-        extends  = interesting_singleton_class_ancestors.reject do |mod|
-          Module != class_of(mod) || are_equal?(mod, singleton_class)
-        end
-
-        add_mixins(tree, prepends.reverse, Trackers::Mixin::Type::Prepend)
-        add_mixins(tree, includes.reverse, Trackers::Mixin::Type::Include)
-        add_mixins(tree, extends.reverse, Trackers::Mixin::Type::Extend)
-      end
-
-      sig do
-        params(
-          tree: RBI::Tree,
-          mods: T::Array[Module],
-          mixin_type: Trackers::Mixin::Type
-        ).void
-      end
-      def add_mixins(tree, mods, mixin_type)
-        mods
-          .select do |mod|
-            name = name_of(mod)
-
-            name && !filtered_mixin?(name)
-          end
-          .map do |mod|
-            name = name_of(mod)
-            push_symbol(name) if name
-
-            qname = qualified_name_of(mod)
-            case mixin_type
-            # TODO: Sorbet currently does not handle prepend
-            # properly for method resolution, so we generate an
-            # include statement instead
-            when Trackers::Mixin::Type::Include, Trackers::Mixin::Type::Prepend
-              tree << RBI::Include.new(T.must(qname))
-            when Trackers::Mixin::Type::Extend
-              tree << RBI::Extend.new(T.must(qname))
-            end
-          end
       end
 
       sig { params(tree: RBI::Tree, name: String, constant: Module).void }
@@ -578,13 +529,6 @@ module Tapioca
         @payload_symbols.include?(symbol_name)
       end
 
-      sig { params(mixin_name: String).returns(T::Boolean) }
-      def filtered_mixin?(mixin_name)
-        # filter T:: namespace mixins that aren't T::Props
-        # T::Props and subconstants have semantic value
-        mixin_name.start_with?("T::") && !mixin_name.start_with?("T::Props")
-      end
-
       SPECIAL_METHOD_NAMES = T.let([
         "!", "~", "+@", "**", "-@", "*", "/", "%", "+", "-", "<<", ">>", "&", "|", "^",
         "<", "<=", "=>", ">", ">=", "==", "===", "!=", "=~", "!~", "<=>", "[]", "[]=", "`",
@@ -657,33 +601,6 @@ module Tapioca
         constant.instance_method(:initialize)
       rescue
         nil
-      end
-
-      sig { params(constant: Module).returns(T::Array[Module]) }
-      def interesting_ancestors_of(constant)
-        inherited_ancestors_ids = Set.new(
-          inherited_ancestors_of(constant).map { |mod| object_id_of(mod) }
-        )
-        # TODO: There is actually a bug here where this will drop modules that
-        # may be included twice. For example:
-        #
-        # ```ruby
-        # class Foo
-        #   prepend Kernel
-        # end
-        # ````
-        # would give:
-        # ```ruby
-        # Foo.ancestors #=> [Kernel, Foo, Object, Kernel, BasicObject]
-        # ````
-        # but since we drop `Kernel` whenever we match it, we would miss
-        # the `prepend Kernel` in the output.
-        #
-        # Instead, we should only drop the tail matches of the ancestors and
-        # inherited ancestors, past the location of the constant itself.
-        constant.ancestors.reject do |mod|
-          inherited_ancestors_ids.include?(object_id_of(mod))
-        end
       end
 
       sig { params(constant: Module).returns(T.nilable(String)) }
