+++ lib/tapioca/internal.rb
@@ -18,6 +18,7 @@ require "tapioca/symbol_table_parser"
 require "tapioca/symbol_loader"
 require "tapioca/compilers/gem/events"
 require "tapioca/compilers/gem/listeners/base"
+require "tapioca/compilers/gem/listeners/mixins"
 require "tapioca/compilers/gem/listeners/sorbet_enums"
 require "tapioca/compilers/gem/listeners/sorbet_helpers"
 require "tapioca/compilers/gem/listeners/sorbet_props"
